Add TestCampaignExecution to manage CardConfigs
This commit adds a TestCampaignExecution class which handles/represents a singele Execution of a complete TestCampaign (containing the according TestElementExecutions) 
This implies that execution of single TestCases within one TestCampaign are no longer supported, as in this situation no direct mapping to the used CardConfig is possible.

Persistence of CardConfigurations and and TestCampaignExcution still needs to be implemented.

git-svn-id: http://subversion.globaltester.org/repo/globaltester/org.globaltester.cardconfiguration.test/trunk@791 cd90bb66-5c64-11e0-9e8e-9b0b1919da2d